fix some stream continuity issues (#2186)

+++ ErsatzTV.Application/Streaming/Queries/FFmpegProcessRequest.cs
@@ -8,4 +8,4 @@ public record FFmpegProcessRequest(
     DateTimeOffset Now,
     bool StartAtZero,
     bool HlsRealtime,
-    long PtsOffset) : IRequest<Either<BaseError, PlayoutItemProcessModel>>;
+    double PtsOffset) : IRequest<Either<BaseError, PlayoutItemProcessModel>>;
